+++ pynest/nest/spatial/__init__.py
@@ -19,4 +19,29 @@
 # You should have received a copy of the GNU General Public License
 # along with NEST.  If not, see <http://www.gnu.org/licenses/>.
 
+import functools as _functools
 from .hl_api_spatial import *  # noqa
+from .hl_api_spatial import DistanceParameter as _DistanceParameter
+
+@_functools.cache
+def __getattr__(name):
+    if name == "distance":
+        return _DistanceParameter()
+    raise AttributeError(f"module {__name__} has no attribute {name}")
+
+# Type annotation to hint at dynamic singleton of DistanceParameter()
+distance: _DistanceParameter
+"""
+A singleton instance representing the distance between two nodes in space.
+
+If used alone, the DistanceObject represents simply the Euclidean
+distance between two nodes.
+
+Alternatively the distance in a single dimension may be chosen. Three
+properties are defined, x, y, and z, which represent the distance in
+their respective dimensions. Note that the distance parameter can only
+be used in contexts with two nodes, e.g. when connecting.
+"""
+
+
+__all__ = [attr for attr in vars().keys() if not attr.startswith("_")]
